Rename formatter to be more generic for all statuses
87-partnerWorkflow
